services/order-service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
psycopg2-binary
sqlalchemy
